Refactor v2 to v1 and adds presentation error message
